[Bugfix] Fix ggml_moe_a8 fake registration dtype for bf16 inputs
The CUDA kernel allocates its output with X's dtype
(csrc/gguf/gguf_kernel.cu ggml_moe_a8: new_empty(..., X.scalar_type()),
dispatched via VLLM_DISPATCH_FLOATING_TYPES), but the fake registration
hardcodes torch.float16. For bf16 GGUF models (supported since #73
removed the Blackwell bf16 restriction), torch.compile and any
FakeTensor-based tracing see fp16 where the kernel produces bf16,
propagating the wrong dtype through the graph. The sibling
_ggml_moe_a8_vec_fake already uses X.dtype.

Return X.dtype in the fake, matching the kernel contract, and add a
FakeTensorMode regression test over fp16 and bf16.

Verified on A100: before the fix the fake returns fp16 for bf16 input;
after, it matches. Both parametrized tests pass against the compiled
extension.

+++ vllm_gguf_plugin/ops.py
@@ -153,8 +153,10 @@ if (
         top_k: torch.SymInt,
         tokens: torch.SymInt,
     ) -> torch.Tensor:
+        # match the CUDA kernel, which allocates its output with X's dtype
+        # (see csrc/gguf/gguf_kernel.cu ggml_moe_a8: new_empty(..., X.scalar_type()))
         return torch.empty(
-            (X.size(0) * top_k, row), dtype=torch.float16, device=W.device
+            (X.size(0) * top_k, row), dtype=X.dtype, device=W.device
         )
